modified lapack_eigen database managing files and added initial advanced search data.

+++ src/Dlighthouse/lighthouse/database/lapack_eigen/dropTables.py
@@ -4,14 +4,26 @@ conn = MySQLdb.connect(host="127.0.0.1", port=3306, user="lighthouse", passwd="y
 
 cursor = conn.cursor()
 
-cursor.execute("DROP TABLE IF EXISTS lighthouse_lapack_eigen")
+x = raw_input('Enter \n\
+              "1" for dropping table lighthouse_lapack_eigen,\n\
+              "2" for lighthouse_lapack_eigen_driver_standard,\n\
+              "3" for lighthouse_lapack_eigen_driver_generalized,\n\
+              "4" for lighthouse_lapack_eigen_computational_standard,\n\
+              "5" for lighthouse_lapack_eigen_computational_generalized, or\n\
+              "all" for dropping all lapack_eigen tables.\n --> ')
 
-cursor.execute("DROP TABLE IF EXISTS lighthouse_lapack_sylvester")
 
-cursor.execute("DROP TABLE IF EXISTS lighthouse_lapack_routineinfo")
+###---------------- for guided search ----------------###
+if x in ["1", "all"]:
+    cursor.execute("DROP TABLE IF EXISTS lighthouse_lapack_eigen")
+
+
+
+###---------------- for advanced search ----------------###
+
+
 
 conn.commit()
-
 cursor.close()
 conn.close()
